Move meta programming to subdirectory
