+++ llarp/handlers/tun.cpp
@@ -13,7 +13,7 @@
 #include <router/abstractrouter.hpp>
 #include <service/context.hpp>
 #include <util/logic.hpp>
-#include <util/memfn.hpp>
+#include <util/meta/memfn.hpp>
 #include <nodedb.hpp>
 
 #include <util/str.hpp>
