+++ llarp/messages/relay_commit.cpp
@@ -12,7 +12,7 @@
 #include <util/buffer.hpp>
 #include <util/logging/logger.hpp>
 #include <util/logic.hpp>
-#include <util/memfn.hpp>
+#include <util/meta/memfn.hpp>
 
 #include <functional>
