+++ llarp/messages/relay_status.cpp
@@ -9,7 +9,7 @@
 #include <util/buffer.hpp>
 #include <util/logging/logger.hpp>
 #include <util/logic.hpp>
-#include <util/memfn.hpp>
+#include <util/meta/memfn.hpp>
 
 #include <functional>
 #include <utility>
