+++ llarp/router/outbound_session_maker.cpp
@@ -6,7 +6,7 @@
 #include <router/i_rc_lookup_handler.hpp>
 #include <link/i_link_manager.hpp>
 #include <util/logic.hpp>
-#include <util/memfn.hpp>
+#include <util/meta/memfn.hpp>
 #include <util/threading.hpp>
 #include <util/status.hpp>
 #include <crypto/crypto.hpp>
