+++ llarp/router/router.cpp
@@ -18,7 +18,7 @@
 #include <util/logging/json_logger.hpp>
 #include <util/logging/logger_syslog.hpp>
 #include <util/logging/logger.hpp>
-#include <util/memfn.hpp>
+#include <util/meta/memfn.hpp>
 #include <util/metrics.hpp>
 #include <util/str.hpp>
 #include <utp/utp.hpp>
