+++ llarp/util/aligned.hpp
@@ -4,8 +4,8 @@
 #include <util/bencode.h>
 #include <util/encode.hpp>
 #include <util/logging/logger.hpp>
+#include <util/meta/traits.hpp>
 #include <util/printer.hpp>
-#include <util/traits.hpp>
 
 #include <array>
 #include <cstddef>
