--- llarp/util/memfn.cpp
@@ -1,1 +1,0 @@
-#include <util/memfn.hpp>
