--- llarp/util/memfn_traits.cpp
@@ -1,1 +1,0 @@
-#include <util/memfn_traits.hpp>
