+++ llarp/util/meta/memfn.cpp
@@ -1,0 +1,1 @@
+#include <util/meta/memfn.hpp>
