+++ llarp/util/meta/object.cpp
@@ -1,0 +1,1 @@
+#include <util/meta/object.hpp>
