--- llarp/util/object.cpp
@@ -1,1 +1,0 @@
-#include <util/object.hpp>
