--- llarp/util/traits.cpp
@@ -1,1 +1,0 @@
-#include <util/traits.hpp>
