--- llarp/util/variant.cpp
@@ -1,1 +1,0 @@
-#include <util/variant.hpp>
