+++ llarp/utp/session.cpp
@@ -4,7 +4,7 @@
 #include <messages/discard.hpp>
 #include <messages/link_intro.hpp>
 #include <util/metrics.hpp>
-#include <util/memfn.hpp>
+#include <util/meta/memfn.hpp>
 
 namespace llarp
 {
