+++ test/util/meta/test_llarp_util_object.cpp
@@ -1,4 +1,4 @@
-#include <util/object.hpp>
+#include <util/meta/object.hpp>
 
 #include <array>
 #include <thread>
