+++ test/util/meta/test_llarp_util_traits.cpp
@@ -1,4 +1,4 @@
-#include <util/traits.hpp>
+#include <util/meta/traits.hpp>
 
 #include <list>
